Extract shared Vite config into vite.config.base.ts
